feat(hero): server-render hero background and defer motion
Move the LCP hero image into a Server Component and load decorative
Framer Motion via dynamic imports so it does not block first paint.
