+++ apps/site/src/components/sections/hero/index.ts
@@ -3,4 +3,5 @@
  * Task: T3.2, T3.3
  */
 export * from './Hero';
+export { HeroBackgroundImage } from './HeroBackgroundImage';
 export { default as HeroWithTiles } from './HeroWithTiles';
